create associated person object on account creation. Closes issue #14
The problem causing this issue was that a new person was created when a
new authorization was created, but the person wasn't associated with
the authorization.account. This caused an exception in de home#show.
Also a new person was created per authorization.

This temporary fix instantiates an empty person, user should be
redirected to a registration form prefilled with omniauth data (if user
created account through omniauth).